Split settings.{h,cpp} god-class along its 6 concerns (closes #1195)
Per the audit in #1194 / #1195, app/entities/settings.{h,cpp} bundled
component data, an entity factory, accessor wrappers, system functions,
JSON serialization, and filesystem I/O in one place — the Acme/God-class
pattern Fabian's chapter explicitly warns about.

This split mirrors the high_score precedent from PR #1230:

| Type / function                          | New home                         |
|------------------------------------------|----------------------------------|
| SettingsState, SettingsPersistence       | components/settings.h            |
| create/destroy_settings_entity()         | entities/settings_entity.{h,cpp} |
| find_/settings_state(), find_/persist()  | systems/settings_system.{h,cpp}  |
| namespace settings { load/save/… }       | systems/settings_system.{h,cpp}  |
| JSON serialization + filesystem helpers  | systems/settings_system.cpp      |

entities/settings.h stays as a back-compat shim re-including the three
new headers (same approach as rhythm.h after #1231, scoring.h after
#1233, rendering.h after #1236). All 22 existing consumers continue to
compile without changes.

test_wasm_beforeunload_lifecycle.cpp patched to read the new
systems/settings_system.cpp path for its persistence-flush assertions.

Validation: 904 Catch2 tests / 2961 assertions pass; zero warnings under
both unity AND non-unity Clang -Wall -Wextra -Werror builds.

Refs #1194, #1203 (relational-normalization parent).
